Tests/LibWeb/Text/input/HTML/document-open-reopened-child-task-delays-parent-load.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    asyncTest(async done => {
        const parentIframe = document.createElement("iframe");
        const initialParentLoad = new Promise(resolve => parentIframe.addEventListener("load", resolve, { once: true }));
        document.body.appendChild(parentIframe);
        await initialParentLoad;

        let parentLoadCount = 0;
        parentIframe.addEventListener("load", () => ++parentLoadCount);

        const parentDocument = parentIframe.contentDocument;
        parentDocument.open().write('<iframe id="child" srcdoc="Initial child document"></iframe>');
        const childIframe = parentDocument.querySelector("#child");
        const initialChildLoad = new Promise(resolve => childIframe.addEventListener("load", resolve, { once: true }));
        await initialChildLoad;

        const childDocument = childIframe.contentDocument;
        const reopenChild = "reopen-child-after-parent-dom-content-loaded";
        const firstCheck = "reopened-child-parent-load-task-check-1";
        const secondCheck = "reopened-child-parent-load-task-check-2";

        parentDocument.addEventListener("DOMContentLoaded", () => window.postMessage(reopenChild, "*"));
        window.addEventListener("message", event => {
            if (event.data === reopenChild) {
                childDocument.open().write("<body>Replacement child document</body>");
                window.postMessage(firstCheck, "*");
                return;
            }
            if (event.data === firstCheck) {
                window.postMessage(secondCheck, "*");
                return;
            }
            if (event.data !== secondCheck)
                return;

            println(`Parent load events before replacement close: ${parentLoadCount}`);
            childDocument.addEventListener("readystatechange", () => {
                if (childDocument.readyState !== "complete")
                    return;
                println("Replacement child reached complete ready state");
                parentIframe.remove();
                done();
            });
            childDocument.close();
        });
        parentDocument.close();
    });
</script>
