Tests/LibWeb/Text/expected/HTML/document-open-reopened-child-task-delays-parent-load.txt

Parent load events before replacement close: 0
Replacement child reached complete ready state
